# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Enactus Platform API
  version: 0.4.0
  description: |
    API de Enactus Platform (eduXaction Colombia).

    **Formato de error único.** Todo error responde con la misma forma:

    ```json
    { "error": { "code": "forbidden", "message": "…", "details": … } }
    ```

    `code` es el identificador estable que el cliente mira para decidir qué
    hacer; `message` es para la persona. Un error inesperado se registra
    entero en el servidor y llega genérico al cliente — nunca un detalle
    interno de PostgreSQL.

    **Dos criterios que conviene tener presentes al leer:**

    - Pedir por id un recurso que existe pero está fuera del alcance del rol
      devuelve **404, no 403**. Un 403 ya confirmaría que ese recurso existe.
    - Un estudiante de Open Learning que llama un endpoint exclusivo de
      Enactus recibe **403, nunca una lista vacía**: una lista vacía haría
      ver un bug de permisos como si fuera "todavía no hay datos".

    **Alcance de lectura.** Ningún listado devuelve todo a todo el mundo: el
    filtro por rol se aplica en el servidor y no se puede cambiar con un
    parámetro de la petición.
servers:
  - url: http://localhost:3000
    description: Desarrollo local
  - url: https://api-staging.eduxaction.co
    description: Staging (pendiente, Fase 6)
  - url: https://api.eduxaction.co
    description: Producción (pendiente, Fase 6)
...
security:
  - bearerAuth: []
paths:
...
  /students/{id}/ruta-progress:
    parameters: [{ $ref: '#/components/parameters/id' }]
    get:
      tags: [students]
      summary: Ruta de Impacto completa
      description: |
        Laboratorios, fases, módulos, objetivos y cursos, con el desbloqueo y
        el estado de cada deadline.

        Un estudiante de Open Learning recibe **403**. Si quien pregunta es
        del equipo pero el estudiante objetivo es de Open Learning, responde
        **409** diciéndolo — nunca una lista vacía.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RutaProgress' }
        '403': { description: Cuenta de Open Learning, o no es su propio progreso }
        '409': { description: El estudiante objetivo es de Open Learning }
  /students/{id}/course-progress/{courseId}:
    parameters:
      - { $ref: '#/components/parameters/id' }
      - { in: path, name: courseId, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [students]
      summary: Progreso en un curso, con las lecciones completadas
      responses:
        '200': { description: OK }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: El estudiante no tiene acceso a ese curso }
  /students/{id}/certificate-eligibility/{laboratoryId}:
    parameters:
      - { $ref: '#/components/parameters/id' }
      - { in: path, name: laboratoryId, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [students]
      summary: Si ya puede emitirse el certificado, y si no, qué falta
      responses: { '200': { description: OK } }
...
  /submissions:
    get:
      tags: [submissions]
      summary: Entregas visibles para el rol
      description: Un estudiante ve **solo las suyas**; no hay parámetro que lo cambie.
      parameters:
        - { $ref: '#/components/parameters/page' }
        - { $ref: '#/components/parameters/pageSize' }
        - { in: query, name: courseId, schema: { type: string, format: uuid } }
        - { in: query, name: studentId, schema: { type: string, format: uuid } }
        - { in: query, name: pending, schema: { type: string, enum: ['true', 'false'] } }
      responses: { '200': { description: Página de entregas } }
    post:
      tags: [submissions]
      summary: Crear una entrega (solo el propio estudiante)
      responses:
        '201': { description: Creada }
        '403': { description: Sin acceso a ese curso o laboratorio }
  /submissions/{id}/grade:
    parameters: [{ $ref: '#/components/parameters/id' }]
    post:
      tags: [submissions]
      summary: Calificar
      description: |
        Requiere `can_grade` **en el contexto del curso**: Open Learning y
        Enactus son permisos separados, y el contexto sale del curso, no del
        cuerpo de la petición.

        La escala se guarda junto a la nota: hoy conviven cuatro escalas en el
        mismo campo y el número por sí solo no significa nada.

        Calificar una actividad **completa su lección** (marca, nunca alterna).
        Queda en `audit_log`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gradingMode]
              properties:
                grade: { type: number, nullable: true }
                gradingMode: { type: string, enum: [points100, passfail, review, scale5] }
                feedback: { type: string }
      responses:
        '200': { description: Calificada }
        '400': { description: Nota fuera del rango de la escala }
        '403': { description: Sin permiso de calificar en ese contexto }
  /submissions/{id}/review:
    parameters: [{ $ref: '#/components/parameters/id' }]
    post:
      tags: [submissions]
      summary: Comentar sin calificar (Mentor)
      description: |
        El Mentor revisa y comenta, pero **no pone nota**. Este endpoint no
        toca `grade` bajo ninguna circunstancia.
      responses: { '200': { description: Revisada } }
  /submissions/{id}:
    parameters: [{ $ref: '#/components/parameters/id' }]
    delete:
      tags: [submissions]
      summary: Borrado lógico
      description: El estudiante solo puede borrar la suya mientras no esté calificada ni comentada.
      responses:
        '204': { description: Borrada }
        '409': { description: Ya calificada o comentada }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token de 12 h. Se renueva con `/auth/refresh`.
  parameters:
    id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
...
    Forbidden:
      description: El rol no tiene permiso
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - validation_error
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - payload_too_large
                - too_many_requests
                - storage_not_configured
                - internal_error
            message: { type: string }
            details: {}
...
    RutaProgress:
      type: object
      properties:
        studentId: { type: string, format: uuid }
        studentName: { type: string }
        laboratories:
          type: array
          items:
            type: object
            properties:
              laboratoryId: { type: string, format: uuid }
              laboratoryName: { type: string }
              phasesTotal: { type: integer }
              phasesDone: { type: integer }
              isComplete: { type: boolean }
              certificateIssued: { type: boolean }
              phases:
                type: array
                items:
                  type: object
                  properties:
                    phaseId: { type: string, format: uuid }
                    orderIndex: { type: integer }
                    title: { type: string }
                    deadline: { type: string, format: date, nullable: true }
                    deadlineStatus:
                      type: string
                      enum: [none, on_track, approaching, overdue]
                      description: Una fase completada nunca aparece atrasada.
                    modulesTotal: { type: integer }
                    modulesDone: { type: integer }
                    isComplete: { type: boolean }
                    isUnlocked: { type: boolean }
                    objectives: { type: array, items: {} }
                    modules: { type: array, items: {} }